iza-os-production/requirements.mcp.txt
# requirements.txt
# Python dependencies for the MCP server

fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
GitPython==3.1.40
neo4j==5.14.1
qdrant-client==1.6.9
openai==1.3.6
anthropic==0.7.8
PyYAML==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
watchdog==3.0.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
tiktoken==0.5.1
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
